backend_django/requirements.txt
# Django Core
Django==5.0.1
django-environ==0.11.2

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# CORS
django-cors-headers==4.3.1

# Google Cloud Storage
google-cloud-storage==2.14.0

# Image Processing
Pillow==10.2.0

# Filtering
django-filter==23.5

# API Documentation
drf-yasg==1.21.7

# Production Server
gunicorn==21.2.0

# Static Files
whitenoise==6.6.0

# Utilities
python-decouple==3.8
